Fixed Jumping Problem
The solution was to set a Default Body Gravity Scale and add that to the first jumpSpeed equation instead of the regular body.gravityScale (since that can change from 2.8 to 17 with a jumpHeight of 4.5). Then, when the velocity.y is less than 0, before you add the absolute value of the current velocity.y, you divide the jumpSpeed by a divider to ensure that it will jump similarily to the first jump when the velocity.y is 0.

There may be a way to have the first jumpSpeed equation keep the body.gravityScale a part of the equation and still have the divider to ensure that the second jump is the same as the first jump.